Add test for PHP lib usage and GH Actions
