+++ phpunit.xml.dist
@@ -4,9 +4,4 @@
 			<directory>tests/</directory>
 		</testsuite>
 	</testsuites>
-	<filter>
-		<whitelist addUncoveredFilesFromWhitelist="true">
-			<directory suffix=".php">src</directory>
-		</whitelist>
-	</filter>
 </phpunit>
