+++ tests/phpunit/NeoWikiTest.php
@@ -2,18 +2,24 @@
 
 declare( strict_types = 1 );
 
-namespace ProfessionalWiki\NeoWiki\Tests;
+namespace ProfessionalWiki\NeoWiki\MediaWiki\Tests;
 
+use OutOfBoundsException;
 use PHPUnit\Framework\TestCase;
-use ProfessionalWiki\NeoWiki\NeoWiki;
+use ProfessionalWiki\NeoWiki\MediaWiki\NeoWikiExtension;
 
 /**
- * @covers \ProfessionalWiki\NeoWiki\NeoWiki
+ * @covers \ProfessionalWiki\NeoWiki\MediaWiki\NeoWikiExtension
  */
 class NeoWikiTest extends TestCase {
 
 	public function testGetInstanceIsSingleton(): void {
-		$this->assertSame( NeoWiki::getInstance(), NeoWiki::getInstance() );
+		$this->assertSame( NeoWikiExtension::getInstance(), NeoWikiExtension::getInstance() );
+	}
+
+	public function testGetServiceFromPhpLibrary(): void {
+		$this->expectException( OutOfBoundsException::class );
+		NeoWikiExtension::getInstance()->getFormatRegistry()->getFormatOrThrow( 'does-not-exist' );
 	}
 
 }
